pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "bookfix"
version = "0.1.0"
dependencies = ["pypdf", "Pillow"]

[project.scripts]
bookfix = "bookfix:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
